change to std input
